Add skipping of copy if file exists
Both copy2 and copytree do not overwrite the files if they are
existing. So, it is necessary to check before copying else a
FileExistsError will be thrown. With this commit, copy will log and
skip the action if file exists.

+++ napi/utils.py
@@ -4,6 +4,7 @@ import os
 import re
 import yaml
 import shutil
+import logging
 import subprocess
 
 
@@ -11,6 +12,11 @@ def copy(src, dst):
     """Copies from source to destination."""
     if not dst:
         raise TypeError("Expected destination path, received None")
+
+    if os.path.exists(dst):
+        logging.warning("Skipping copy as file exists")
+        return
+
     os.makedirs(os.path.dirname(dst), exist_ok=True)
     if os.path.isdir(src):
         shutil.copytree(src, dst)
